refactor: :art: rebuild the CD system

+++ logic/Preparation/Utility/GameData.cs
@@ -8,7 +8,7 @@ namespace Preparation.Utility
         #region 基本常数
         public const int numOfStepPerSecond = 20;     // 每秒行走的步数
         public const int frameDuration = 50;         // 每帧时长
-        public const long checkInterval = 50;  // 检查位置标志、补充子弹的帧时长
+        public const int checkInterval = 50;  // 检查位置标志、补充子弹的帧时长
 
         public const long gameDuration = 600000;      // 游戏时长600000ms = 10min
